refactor: move json explorer widget into page feature
