+++ src/pages/json-explorer/PageJsonExplorer.tsx
@@ -1,4 +1,4 @@
-import { JsonExplorer } from './widgets/JsonExplorer/JsonExplorer';
+import { JsonExplorer } from './features/JsonExplorer/ui/JsonExplorer';
 
 export function PageJsonExplorer() {
   return <JsonExplorer />;
