+++ src/pages/json-explorer/features/JsonExplorer/lib/extractKeyPaths.ts
