+++ src/pages/json-explorer/features/JsonExplorer/ui/CustomJsonModal.tsx
