+++ src/pages/json-explorer/features/JsonExplorer/ui/DataTable.tsx
