+++ src/pages/json-explorer/features/JsonExplorer/ui/JsonDetailsPanel.tsx
