+++ src/pages/json-explorer/features/JsonExplorer/ui/SearchBar.tsx
